Setters and getters moved into dedicated namespaces
