+++ StoryLine.Selenium.Runner/Google/Mappings/SearchModelMapping.cs
@@ -1,7 +1,8 @@
-﻿using StoryLine.Selenium.Bindings;
+﻿using StoryLine.Selenium.Getters;
 using StoryLine.Selenium.Mappings;
 using StoryLine.Selenium.Runner.Google.Models;
 using StoryLine.Selenium.Runner.Google.Pages;
+using StoryLine.Selenium.Setters;
 
 namespace StoryLine.Selenium.Runner.Google.Mappings
 {
