+++ StoryLine.Selenium.Runner/RealtOnliner/Mappings/SearchModelMapping.cs
@@ -1,7 +1,8 @@
-﻿using StoryLine.Selenium.Bindings;
+﻿using StoryLine.Selenium.Getters;
 using StoryLine.Selenium.Mappings;
 using StoryLine.Selenium.Runner.RealtOnliner.Models;
 using StoryLine.Selenium.Runner.RealtOnliner.Pages;
+using StoryLine.Selenium.Setters;
 
 namespace StoryLine.Selenium.Runner.RealtOnliner.Mappings
 {
