+++ StoryLine.Selenium/Setters/ICollectionValueSetter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using OpenQA.Selenium;
 
-namespace StoryLine.Selenium.Bindings
+namespace StoryLine.Selenium.Setters
 {
     public interface ICollectionValueSetter
     {
